Generate a attendee pin for the moderator when enabling SIP and return it
